chore: move ImagePreview to components
